test(app): isolate plugin icon regression setup
